InvalidRectError - floating point error Issue #152 (#153)

+++ lib/src/logic/cropper/image_image_cropper.dart
@@ -1,8 +1,6 @@
 import 'dart:async';
 import 'dart:math';
-
 import 'dart:typed_data';
-
 import 'dart:ui';
 
 import 'package:crop_your_image/src/logic/cropper/errors.dart';
@@ -24,10 +22,10 @@ class ImageImageCropper extends ImageCropper<Image> {
     ImageFormat outputFormat = ImageFormat.jpeg,
     ImageShape shape = ImageShape.rectangle,
   }) {
-    if (topLeft.dx.isNegative ||
-        topLeft.dy.isNegative ||
-        bottomRight.dx.isNegative ||
-        bottomRight.dy.isNegative ||
+    if (topLeft.dx.toInt().isNegative ||
+        topLeft.dy.toInt().isNegative ||
+        bottomRight.dx.toInt().isNegative ||
+        bottomRight.dy.toInt().isNegative ||
         topLeft.dx.toInt() > original.width ||
         topLeft.dy.toInt() > original.height ||
         bottomRight.dx.toInt() > original.width ||
